fix: address PR review comments and split oversized route modules
Review fixes (PR#366):
- Critical: publish EventBridge event BEFORE updating to IN_PROGRESS to prevent
  tenants getting stuck if the follow-up DynamoDB update also fails
- Security: restrict provisionedResources (IAM ARNs, S3, CloudWatch) to
  PLATFORM_ADMIN role only in GET /api/tenants/:id/provision
- Fix provisionedAt to return undefined instead of null for consistency
- Improve EventBridge error message with context prefix for easier debugging
- Fix TenantData test interface to use typed string-literal unions
- Remove empty-ID tenant test that validated an impossible production state

Module refactoring (quality gate):
- Split index.ts (1097→226 lines) into routes/settings.ts, routes/tenants.ts,
  routes/provisioning.ts and routes/shared.ts
- Split index.test.ts (1549→330 lines) into __tests__/settings.test.ts,
  __tests__/tenants.test.ts, __tests__/tenants-tier.test.ts,
  __tests__/provisioning.test.ts
- All files now under 600-line quality gate threshold

Also includes vote page UI refresh (Tailwind, per-candidate state, live board)
and corresponding test updates to use getParticipantTeams API mock.
