+++ backend/services/control-plane/tenant-management/src/index.test.ts
@@ -327,1227 +327,4 @@ describe('テナント管理API', () => {
       });
     });
   });
-
-  describe('POST /api/tenants', () => {
-    it('有効なデータでテナントを作成できるべき', async () => {
-      const mockTenant = createMockTenant();
-      mockTenantRepoFunctions.findBySlug.mockResolvedValue(null);
-      mockTenantRepoFunctions.create.mockResolvedValue(mockTenant);
-
-      const res = await app.request('/api/tenants', {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify({
-          name: 'Test Organization',
-          slug: 'test-organization',
-          adminEmail: 'admin@test.com',
-          tier: 'FREE',
-          status: 'ACTIVE',
-        }),
-      });
-
-      expect(res.status).toBe(201);
-      const body = await res.json();
-      expect(body).toMatchObject({
-        name: mockTenant.name,
-        adminEmail: mockTenant.adminEmail,
-        tier: mockTenant.tier,
-        status: mockTenant.status,
-      });
-      expect(body.id).toBeDefined();
-    });
-
-    it('デフォルト値でテナントを作成できるべき', async () => {
-      const mockTenant = createMockTenant({
-        slug: 'default-tenant',
-        name: 'Default Tenant',
-        adminEmail: 'default@test.com',
-      });
-      mockTenantRepoFunctions.findBySlug.mockResolvedValue(null);
-      mockTenantRepoFunctions.create.mockResolvedValue(mockTenant);
-
-      const res = await app.request('/api/tenants', {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify({
-          name: 'Default Tenant',
-          slug: 'default-tenant',
-          adminEmail: 'default@test.com',
-        }),
-      });
-
-      expect(res.status).toBe(201);
-      const body = await res.json();
-      expect(body.tier).toBe('FREE');
-      expect(body.status).toBe('ACTIVE');
-    });
-
-    it('不正なメールアドレスでバリデーションエラーになるべき', async () => {
-      const res = await app.request('/api/tenants', {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify({
-          name: 'Invalid Email Tenant',
-          slug: 'invalid-email-tenant',
-          adminEmail: 'not-an-email',
-        }),
-      });
-
-      expect(res.status).toBe(400);
-      const body = await res.json();
-      expect(body.error).toBe('Validation error');
-      expect(body.details).toBeDefined();
-    });
-
-    it('空の名前でバリデーションエラーになるべき', async () => {
-      const res = await app.request('/api/tenants', {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify({
-          name: '',
-          slug: 'empty-name-tenant',
-          adminEmail: 'valid@test.com',
-        }),
-      });
-
-      expect(res.status).toBe(400);
-      const body = await res.json();
-      expect(body.error).toBe('Validation error');
-    });
-
-    it('重複したスラッグで409エラーになるべき', async () => {
-      const existingTenant = createMockTenant({ slug: 'duplicate-tenant' });
-      mockTenantRepoFunctions.findBySlug.mockResolvedValue(existingTenant);
-
-      const res = await app.request('/api/tenants', {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify({
-          name: 'Duplicate Slug Tenant',
-          slug: 'duplicate-tenant',
-          adminEmail: 'duplicate@test.com',
-        }),
-      });
-
-      expect(res.status).toBe(409);
-      const body = await res.json();
-      expect(body.error).toContain('already exists');
-    });
-
-    it('不正なtier値でバリデーションエラーになるべき', async () => {
-      const res = await app.request('/api/tenants', {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify({
-          name: 'Invalid Tier',
-          slug: 'invalid-tier-tenant',
-          adminEmail: 'tier@test.com',
-          tier: 'INVALID_TIER',
-        }),
-      });
-
-      expect(res.status).toBe(400);
-      const body = await res.json();
-      expect(body.error).toBe('Validation error');
-    });
-  });
-
-  describe('GET /api/tenants', () => {
-    it('すべてのテナントを取得できるべき', async () => {
-      const mockTenants = [
-        createMockTenant({
-          id: '01HJXK5K3VDXK5YPNZBKRT5001',
-          name: 'Tenant 1',
-        }),
-        createMockTenant({
-          id: '01HJXK5K3VDXK5YPNZBKRT5002',
-          name: 'Tenant 2',
-        }),
-        createMockTenant({
-          id: '01HJXK5K3VDXK5YPNZBKRT5003',
-          name: 'Tenant 3',
-        }),
-      ];
-      mockTenantRepoFunctions.count.mockResolvedValue(3);
-      mockTenantRepoFunctions.list.mockResolvedValue({
-        tenants: mockTenants,
-        lastKey: undefined,
-      });
-
-      const res = await app.request('/api/tenants');
-      expect(res.status).toBe(200);
-      const body = await res.json();
-      expect(body.data).toBeDefined();
-      expect(Array.isArray(body.data)).toBe(true);
-      expect(body.data.length).toBe(3);
-      expect(body.pagination).toBeDefined();
-      expect(body.pagination.total).toBe(3);
-    });
-
-    it('テナントが存在しない場合は空配列を返すべき', async () => {
-      mockTenantRepoFunctions.count.mockResolvedValue(0);
-      mockTenantRepoFunctions.list.mockResolvedValue({
-        tenants: [],
-        lastKey: undefined,
-      });
-
-      const res = await app.request('/api/tenants');
-      expect(res.status).toBe(200);
-      const body = await res.json();
-      expect(body.data).toEqual([]);
-      expect(body.pagination.total).toBe(0);
-    });
-
-    it('ページネーション: limitパラメータが正しく動作するべき', async () => {
-      const mockTenants = [
-        createMockTenant({
-          id: '01HJXK5K3VDXK5YPNZBKRT5001',
-          name: 'Tenant 1',
-        }),
-        createMockTenant({
-          id: '01HJXK5K3VDXK5YPNZBKRT5002',
-          name: 'Tenant 2',
-        }),
-      ];
-      mockTenantRepoFunctions.count.mockResolvedValue(3);
-      mockTenantRepoFunctions.list.mockResolvedValue({
-        tenants: mockTenants,
-        lastKey: { PK: 'TENANT#01HJXK5K3VDXK5YPNZBKRT5002' },
-      });
-
-      const res = await app.request('/api/tenants?limit=2');
-      expect(res.status).toBe(200);
-      const body = await res.json();
-      expect(body.data.length).toBeLessThanOrEqual(2);
-      expect(body.pagination.limit).toBe(2);
-      expect(body.pagination.hasNextPage).toBe(true);
-    });
-
-    it('ページネーション: limitは100に制限されるべき', async () => {
-      mockTenantRepoFunctions.count.mockResolvedValue(0);
-      mockTenantRepoFunctions.list.mockResolvedValue({
-        tenants: [],
-        lastKey: undefined,
-      });
-
-      const res = await app.request('/api/tenants?limit=999999');
-      expect(res.status).toBe(200);
-      const body = await res.json();
-      expect(body.pagination.limit).toBe(100);
-    });
-
-    it('ページネーション: ゼロのlimit値は1にフォールバックされるべき', async () => {
-      mockTenantRepoFunctions.count.mockResolvedValue(0);
-      mockTenantRepoFunctions.list.mockResolvedValue({
-        tenants: [],
-        lastKey: undefined,
-      });
-
-      const res = await app.request('/api/tenants?limit=0');
-      expect(res.status).toBe(200);
-      const body = await res.json();
-      expect(body.pagination.limit).toBe(1);
-    });
-  });
-
-  describe('GET /api/tenants/:id', () => {
-    it('IDでテナントを取得できるべき', async () => {
-      const mockTenant = createMockTenant();
-      mockTenantRepoFunctions.findById.mockResolvedValue(mockTenant);
-
-      const res = await app.request(`/api/tenants/${mockTenant.id}`);
-      expect(res.status).toBe(200);
-      const body = await res.json();
-      expect(body).toMatchObject({
-        id: mockTenant.id,
-        name: mockTenant.name,
-        adminEmail: mockTenant.adminEmail,
-      });
-    });
-
-    it('存在しないIDで404エラーになるべき', async () => {
-      mockTenantRepoFunctions.findById.mockResolvedValue(null);
-
-      const nonExistentId = '01HJXK5K3VDXK5YPNZBKRT5XYZ';
-      const res = await app.request(`/api/tenants/${nonExistentId}`);
-      expect(res.status).toBe(404);
-      const body = await res.json();
-      expect(body.error).toBe('Tenant not found');
-    });
-
-    it('不正なID形式で400エラーになるべき', async () => {
-      const res = await app.request('/api/tenants/invalid@id!');
-      expect(res.status).toBe(400);
-      const body = await res.json();
-      expect(body.error).toBe('Invalid tenant ID');
-    });
-  });
-
-  describe('PATCH /api/tenants/:id', () => {
-    it('テナント名を更新できるべき', async () => {
-      const mockTenant = createMockTenant();
-      const updatedTenant = { ...mockTenant, name: 'Updated Name' };
-      mockTenantRepoFunctions.update.mockResolvedValue(updatedTenant);
-
-      const res = await app.request(`/api/tenants/${mockTenant.id}`, {
-        method: 'PATCH',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify({ name: 'Updated Name' }),
-      });
-
-      expect(res.status).toBe(200);
-      const body = await res.json();
-      expect(body.name).toBe('Updated Name');
-    });
-
-    it('テナントのステータスを更新できるべき', async () => {
-      const mockTenant = createMockTenant();
-      const updatedTenant = { ...mockTenant, status: 'SUSPENDED' as const };
-      mockTenantRepoFunctions.update.mockResolvedValue(updatedTenant);
-
-      const res = await app.request(`/api/tenants/${mockTenant.id}`, {
-        method: 'PATCH',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify({ status: 'SUSPENDED' }),
-      });
-
-      expect(res.status).toBe(200);
-      const body = await res.json();
-      expect(body.status).toBe('SUSPENDED');
-    });
-
-    it('テナントのtierを更新できるべき', async () => {
-      const mockTenant = createMockTenant();
-      const updatedTenant = {
-        ...mockTenant,
-        tier: 'ENTERPRISE' as const,
-        isolationModel: 'SILO' as const,
-        provisioningStatus: 'PENDING' as const,
-      };
-      mockTenantRepoFunctions.findById.mockResolvedValue(mockTenant);
-      mockTenantRepoFunctions.update.mockResolvedValue(updatedTenant);
-
-      const res = await app.request(`/api/tenants/${mockTenant.id}`, {
-        method: 'PATCH',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify({ tier: 'ENTERPRISE' }),
-      });
-
-      expect(res.status).toBe(200);
-      const body = await res.json();
-      expect(body.tier).toBe('ENTERPRISE');
-    });
-
-    it('複数のフィールドを同時に更新できるべき', async () => {
-      const mockTenant = createMockTenant();
-      const updatedTenant = {
-        ...mockTenant,
-        name: 'New Name',
-        tier: 'PRO' as const,
-        status: 'ACTIVE' as const,
-      };
-      // tier変更が含まれる場合はfindByIdが呼ばれる
-      mockTenantRepoFunctions.findById.mockResolvedValue(mockTenant);
-      mockTenantRepoFunctions.update.mockResolvedValue(updatedTenant);
-
-      const res = await app.request(`/api/tenants/${mockTenant.id}`, {
-        method: 'PATCH',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify({
-          name: 'New Name',
-          tier: 'PRO',
-          status: 'ACTIVE',
-        }),
-      });
-
-      expect(res.status).toBe(200);
-      const body = await res.json();
-      expect(body.name).toBe('New Name');
-      expect(body.tier).toBe('PRO');
-      expect(body.status).toBe('ACTIVE');
-    });
-
-    it('存在しないIDで404エラーになるべき', async () => {
-      const { ConditionalCheckFailedException } =
-        await import('@aws-sdk/client-dynamodb');
-      mockTenantRepoFunctions.update.mockRejectedValue(
-        new ConditionalCheckFailedException({
-          $metadata: {},
-          message: 'Condition check failed',
-        })
-      );
-
-      const nonExistentId = '01HJXK5K3VDXK5YPNZBKRT5XYZ';
-      const res = await app.request(`/api/tenants/${nonExistentId}`, {
-        method: 'PATCH',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify({ name: 'Updated Name' }),
-      });
-
-      expect(res.status).toBe(404);
-      const body = await res.json();
-      expect(body.error).toBe('Tenant not found');
-    });
-
-    it('不正なID形式で400エラーになるべき', async () => {
-      const res = await app.request('/api/tenants/invalid@id!', {
-        method: 'PATCH',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify({ name: 'Updated Name' }),
-      });
-
-      expect(res.status).toBe(400);
-      const body = await res.json();
-      expect(body.error).toBe('Invalid tenant ID');
-    });
-
-    describe('tier変更ビジネスロジック', () => {
-      it('FREE→PRO変更時はisolationModelが変わらないためre-provisioningが発生しないべき', async () => {
-        const mockTenant = createMockTenant({
-          tier: 'FREE',
-          isolationModel: 'POOL',
-          provisioningStatus: 'COMPLETED',
-        });
-        const updatedTenant = { ...mockTenant, tier: 'PRO' as const };
-        mockTenantRepoFunctions.findById.mockResolvedValue(mockTenant);
-        mockTenantRepoFunctions.update.mockResolvedValue(updatedTenant);
-
-        const res = await app.request(`/api/tenants/${mockTenant.id}`, {
-          method: 'PATCH',
-          headers: { 'Content-Type': 'application/json' },
-          body: JSON.stringify({ tier: 'PRO' }),
-        });
-
-        expect(res.status).toBe(200);
-        const body = await res.json();
-        expect(body.tier).toBe('PRO');
-        // update should NOT include isolationModel or provisioningStatus changes
-        expect(mockTenantRepoFunctions.update).toHaveBeenCalledWith(
-          mockTenant.id,
-          expect.not.objectContaining({
-            isolationModel: expect.anything(),
-            provisioningStatus: expect.anything(),
-          })
-        );
-      });
-
-      it('PRO→FREE変更時はisolationModelが変わらないためre-provisioningが発生しないべき', async () => {
-        const mockTenant = createMockTenant({
-          tier: 'PRO',
-          isolationModel: 'POOL',
-          provisioningStatus: 'COMPLETED',
-        });
-        const updatedTenant = { ...mockTenant, tier: 'FREE' as const };
-        mockTenantRepoFunctions.findById.mockResolvedValue(mockTenant);
-        mockTenantRepoFunctions.update.mockResolvedValue(updatedTenant);
-
-        const res = await app.request(`/api/tenants/${mockTenant.id}`, {
-          method: 'PATCH',
-          headers: { 'Content-Type': 'application/json' },
-          body: JSON.stringify({ tier: 'FREE' }),
-        });
-
-        expect(res.status).toBe(200);
-        const body = await res.json();
-        expect(body.tier).toBe('FREE');
-        expect(mockTenantRepoFunctions.update).toHaveBeenCalledWith(
-          mockTenant.id,
-          expect.not.objectContaining({
-            isolationModel: expect.anything(),
-            provisioningStatus: expect.anything(),
-          })
-        );
-      });
-
-      it('FREE→ENTERPRISE変更時はPOOL→SILOに変わりre-provisioningが発生すべき', async () => {
-        const mockTenant = createMockTenant({
-          tier: 'FREE',
-          isolationModel: 'POOL',
-          provisioningStatus: 'COMPLETED',
-        });
-        const updatedTenant = {
-          ...mockTenant,
-          tier: 'ENTERPRISE' as const,
-          isolationModel: 'SILO' as const,
-          provisioningStatus: 'PENDING' as const,
-        };
-        mockTenantRepoFunctions.findById.mockResolvedValue(mockTenant);
-        mockTenantRepoFunctions.update.mockResolvedValue(updatedTenant);
-
-        const res = await app.request(`/api/tenants/${mockTenant.id}`, {
-          method: 'PATCH',
-          headers: { 'Content-Type': 'application/json' },
-          body: JSON.stringify({ tier: 'ENTERPRISE' }),
-        });
-
-        expect(res.status).toBe(200);
-        const body = await res.json();
-        expect(body.tier).toBe('ENTERPRISE');
-        expect(body.isolationModel).toBe('SILO');
-        expect(body.provisioningStatus).toBe('PENDING');
-        expect(mockTenantRepoFunctions.update).toHaveBeenCalledWith(
-          mockTenant.id,
-          {
-            tier: 'ENTERPRISE',
-            isolationModel: 'SILO',
-            provisioningStatus: 'PENDING',
-          }
-        );
-      });
-
-      it('PRO→ENTERPRISE変更時はPOOL→SILOに変わりre-provisioningが発生すべき', async () => {
-        const mockTenant = createMockTenant({
-          tier: 'PRO',
-          isolationModel: 'POOL',
-          provisioningStatus: 'COMPLETED',
-        });
-        const updatedTenant = {
-          ...mockTenant,
-          tier: 'ENTERPRISE' as const,
-          isolationModel: 'SILO' as const,
-          provisioningStatus: 'PENDING' as const,
-        };
-        mockTenantRepoFunctions.findById.mockResolvedValue(mockTenant);
-        mockTenantRepoFunctions.update.mockResolvedValue(updatedTenant);
-
-        const res = await app.request(`/api/tenants/${mockTenant.id}`, {
-          method: 'PATCH',
-          headers: { 'Content-Type': 'application/json' },
-          body: JSON.stringify({ tier: 'ENTERPRISE' }),
-        });
-
-        expect(res.status).toBe(200);
-        const body = await res.json();
-        expect(body.tier).toBe('ENTERPRISE');
-        expect(body.isolationModel).toBe('SILO');
-        expect(body.provisioningStatus).toBe('PENDING');
-        expect(mockTenantRepoFunctions.update).toHaveBeenCalledWith(
-          mockTenant.id,
-          {
-            tier: 'ENTERPRISE',
-            isolationModel: 'SILO',
-            provisioningStatus: 'PENDING',
-          }
-        );
-      });
-
-      it('ENTERPRISE→FREE変更時はSILO→POOLに変わりre-provisioningが発生すべき', async () => {
-        const mockTenant = createMockTenant({
-          tier: 'ENTERPRISE',
-          isolationModel: 'SILO',
-          provisioningStatus: 'COMPLETED',
-        });
-        const updatedTenant = {
-          ...mockTenant,
-          tier: 'FREE' as const,
-          isolationModel: 'POOL' as const,
-          provisioningStatus: 'PENDING' as const,
-        };
-        mockTenantRepoFunctions.findById.mockResolvedValue(mockTenant);
-        mockTenantRepoFunctions.update.mockResolvedValue(updatedTenant);
-
-        const res = await app.request(`/api/tenants/${mockTenant.id}`, {
-          method: 'PATCH',
-          headers: { 'Content-Type': 'application/json' },
-          body: JSON.stringify({ tier: 'FREE' }),
-        });
-
-        expect(res.status).toBe(200);
-        const body = await res.json();
-        expect(body.tier).toBe('FREE');
-        expect(body.isolationModel).toBe('POOL');
-        expect(body.provisioningStatus).toBe('PENDING');
-        expect(mockTenantRepoFunctions.update).toHaveBeenCalledWith(
-          mockTenant.id,
-          {
-            tier: 'FREE',
-            isolationModel: 'POOL',
-            provisioningStatus: 'PENDING',
-          }
-        );
-      });
-
-      it('ENTERPRISE→PRO変更時はSILO→POOLに変わりre-provisioningが発生すべき', async () => {
-        const mockTenant = createMockTenant({
-          tier: 'ENTERPRISE',
-          isolationModel: 'SILO',
-          provisioningStatus: 'COMPLETED',
-        });
-        const updatedTenant = {
-          ...mockTenant,
-          tier: 'PRO' as const,
-          isolationModel: 'POOL' as const,
-          provisioningStatus: 'PENDING' as const,
-        };
-        mockTenantRepoFunctions.findById.mockResolvedValue(mockTenant);
-        mockTenantRepoFunctions.update.mockResolvedValue(updatedTenant);
-
-        const res = await app.request(`/api/tenants/${mockTenant.id}`, {
-          method: 'PATCH',
-          headers: { 'Content-Type': 'application/json' },
-          body: JSON.stringify({ tier: 'PRO' }),
-        });
-
-        expect(res.status).toBe(200);
-        const body = await res.json();
-        expect(body.tier).toBe('PRO');
-        expect(body.isolationModel).toBe('POOL');
-        expect(body.provisioningStatus).toBe('PENDING');
-        expect(mockTenantRepoFunctions.update).toHaveBeenCalledWith(
-          mockTenant.id,
-          {
-            tier: 'PRO',
-            isolationModel: 'POOL',
-            provisioningStatus: 'PENDING',
-          }
-        );
-      });
-
-      it('tier変更時に存在しないテナントは404エラーになるべき', async () => {
-        mockTenantRepoFunctions.findById.mockResolvedValue(null);
-
-        const nonExistentId = '01HJXK5K3VDXK5YPNZBKRT5XYZ';
-        const res = await app.request(`/api/tenants/${nonExistentId}`, {
-          method: 'PATCH',
-          headers: { 'Content-Type': 'application/json' },
-          body: JSON.stringify({ tier: 'ENTERPRISE' }),
-        });
-
-        expect(res.status).toBe(404);
-        const body = await res.json();
-        expect(body.error).toBe('Tenant not found');
-        expect(mockTenantRepoFunctions.update).not.toHaveBeenCalled();
-      });
-
-      it('tier変更とname変更を同時に行えるべき', async () => {
-        const mockTenant = createMockTenant({
-          tier: 'FREE',
-          isolationModel: 'POOL',
-          provisioningStatus: 'COMPLETED',
-        });
-        const updatedTenant = {
-          ...mockTenant,
-          name: 'Updated Enterprise Tenant',
-          tier: 'ENTERPRISE' as const,
-          isolationModel: 'SILO' as const,
-          provisioningStatus: 'PENDING' as const,
-        };
-        mockTenantRepoFunctions.findById.mockResolvedValue(mockTenant);
-        mockTenantRepoFunctions.update.mockResolvedValue(updatedTenant);
-
-        const res = await app.request(`/api/tenants/${mockTenant.id}`, {
-          method: 'PATCH',
-          headers: { 'Content-Type': 'application/json' },
-          body: JSON.stringify({
-            name: 'Updated Enterprise Tenant',
-            tier: 'ENTERPRISE',
-          }),
-        });
-
-        expect(res.status).toBe(200);
-        const body = await res.json();
-        expect(body.name).toBe('Updated Enterprise Tenant');
-        expect(body.tier).toBe('ENTERPRISE');
-        expect(body.isolationModel).toBe('SILO');
-        expect(body.provisioningStatus).toBe('PENDING');
-        expect(mockTenantRepoFunctions.update).toHaveBeenCalledWith(
-          mockTenant.id,
-          {
-            name: 'Updated Enterprise Tenant',
-            tier: 'ENTERPRISE',
-            isolationModel: 'SILO',
-            provisioningStatus: 'PENDING',
-          }
-        );
-      });
-    });
-  });
-
-  describe('DELETE /api/tenants/:id', () => {
-    it('テナントを削除できるべき', async () => {
-      const mockTenant = createMockTenant();
-      mockTenantRepoFunctions.findById.mockResolvedValue(mockTenant);
-      mockTenantRepoFunctions.delete.mockResolvedValue(undefined);
-
-      const res = await app.request(`/api/tenants/${mockTenant.id}`, {
-        method: 'DELETE',
-      });
-
-      expect(res.status).toBe(200);
-      const body = await res.json();
-      expect(body).toEqual({
-        success: true,
-        message: 'Tenant deleted successfully',
-      });
-    });
-
-    it('存在しないIDで404エラーになるべき', async () => {
-      mockTenantRepoFunctions.findById.mockResolvedValue(null);
-
-      const nonExistentId = '01HJXK5K3VDXK5YPNZBKRT5XYZ';
-      const res = await app.request(`/api/tenants/${nonExistentId}`, {
-        method: 'DELETE',
-      });
-
-      expect(res.status).toBe(404);
-      const body = await res.json();
-      expect(body.error).toBe('Tenant not found');
-    });
-
-    it('不正なID形式で400エラーになるべき', async () => {
-      const res = await app.request('/api/tenants/invalid@id!', {
-        method: 'DELETE',
-      });
-
-      expect(res.status).toBe(400);
-      const body = await res.json();
-      expect(body.error).toBe('Invalid tenant ID');
-    });
-
-    it('削除後に同じIDで再度削除すると404エラーになるべき', async () => {
-      const mockTenant = createMockTenant();
-
-      // 最初の削除は成功
-      mockTenantRepoFunctions.findById.mockResolvedValueOnce(mockTenant);
-      mockTenantRepoFunctions.delete.mockResolvedValueOnce(undefined);
-
-      await app.request(`/api/tenants/${mockTenant.id}`, {
-        method: 'DELETE',
-      });
-
-      // 2回目の削除は404
-      mockTenantRepoFunctions.findById.mockResolvedValueOnce(null);
-
-      const res = await app.request(`/api/tenants/${mockTenant.id}`, {
-        method: 'DELETE',
-      });
-
-      expect(res.status).toBe(404);
-      const body = await res.json();
-      expect(body.error).toBe('Tenant not found');
-    });
-  });
-
-  describe('GET /api/settings', () => {
-    it('デフォルト設定を返すべき（設定が未保存の場合）', async () => {
-      mockSettingRepoFunctions.get.mockResolvedValue(null);
-
-      const res = await app.request('/api/settings');
-      expect(res.status).toBe(200);
-      const body = await res.json();
-      expect(body).toEqual({
-        platform: {
-          platformName: 'TenkaCloud',
-          language: 'ja',
-          timezone: 'Asia/Tokyo',
-        },
-        security: {
-          mfaRequired: false,
-          sessionTimeoutMinutes: 60,
-          maxLoginAttempts: 5,
-        },
-        notifications: {
-          emailNotificationsEnabled: true,
-          systemAlertsEnabled: true,
-          maintenanceNotificationsEnabled: true,
-        },
-        appearance: {
-          theme: 'system',
-        },
-      });
-    });
-
-    it('保存された設定を返すべき', async () => {
-      const savedSettings = {
-        platform: {
-          platformName: 'Custom Platform',
-          language: 'en',
-          timezone: 'UTC',
-        },
-        security: {
-          mfaRequired: true,
-          sessionTimeoutMinutes: 30,
-          maxLoginAttempts: 3,
-        },
-        notifications: {
-          emailNotificationsEnabled: false,
-          systemAlertsEnabled: true,
-          maintenanceNotificationsEnabled: false,
-        },
-        appearance: { theme: 'dark' },
-      };
-
-      mockSettingRepoFunctions.get.mockImplementation(async (key: string) => {
-        const categoryMap: Record<string, unknown> = {
-          platform: savedSettings.platform,
-          security: savedSettings.security,
-          notifications: savedSettings.notifications,
-          appearance: savedSettings.appearance,
-        };
-        return categoryMap[key]
-          ? { key, value: JSON.stringify(categoryMap[key]), category: 'system' }
-          : null;
-      });
-
-      const res = await app.request('/api/settings');
-      expect(res.status).toBe(200);
-      const body = await res.json();
-      expect(body).toEqual(savedSettings);
-    });
-  });
-
-  describe('PUT /api/settings', () => {
-    it('設定を保存できるべき', async () => {
-      mockSettingRepoFunctions.set.mockResolvedValue({});
-
-      const newSettings = {
-        platform: {
-          platformName: 'My Platform',
-          language: 'en',
-          timezone: 'America/New_York',
-        },
-        security: {
-          mfaRequired: true,
-          sessionTimeoutMinutes: 120,
-          maxLoginAttempts: 10,
-        },
-        notifications: {
-          emailNotificationsEnabled: true,
-          systemAlertsEnabled: false,
-          maintenanceNotificationsEnabled: true,
-        },
-        appearance: {
-          theme: 'dark',
-        },
-      };
-
-      const res = await app.request('/api/settings', {
-        method: 'PUT',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify(newSettings),
-      });
-
-      expect(res.status).toBe(200);
-      const body = await res.json();
-      expect(body).toEqual({
-        success: true,
-        message: 'Settings saved successfully',
-      });
-      expect(mockSettingRepoFunctions.set).toHaveBeenCalledTimes(4);
-    });
-
-    it('不正な言語設定でバリデーションエラーになるべき', async () => {
-      const invalidSettings = {
-        platform: {
-          platformName: 'Test',
-          language: 'invalid',
-          timezone: 'Asia/Tokyo',
-        },
-        security: {
-          mfaRequired: false,
-          sessionTimeoutMinutes: 60,
-          maxLoginAttempts: 5,
-        },
-        notifications: {
-          emailNotificationsEnabled: true,
-          systemAlertsEnabled: true,
-          maintenanceNotificationsEnabled: true,
-        },
-        appearance: {
-          theme: 'system',
-        },
-      };
-
-      const res = await app.request('/api/settings', {
-        method: 'PUT',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify(invalidSettings),
-      });
-
-      expect(res.status).toBe(400);
-      const body = await res.json();
-      expect(body.error).toBe('Validation error');
-    });
-
-    it('不正なテーマ設定でバリデーションエラーになるべき', async () => {
-      const invalidSettings = {
-        platform: {
-          platformName: 'Test',
-          language: 'ja',
-          timezone: 'Asia/Tokyo',
-        },
-        security: {
-          mfaRequired: false,
-          sessionTimeoutMinutes: 60,
-          maxLoginAttempts: 5,
-        },
-        notifications: {
-          emailNotificationsEnabled: true,
-          systemAlertsEnabled: true,
-          maintenanceNotificationsEnabled: true,
-        },
-        appearance: {
-          theme: 'invalid-theme',
-        },
-      };
-
-      const res = await app.request('/api/settings', {
-        method: 'PUT',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify(invalidSettings),
-      });
-
-      expect(res.status).toBe(400);
-      const body = await res.json();
-      expect(body.error).toBe('Validation error');
-    });
-
-    it('セッションタイムアウトが範囲外でバリデーションエラーになるべき', async () => {
-      const invalidSettings = {
-        platform: {
-          platformName: 'Test',
-          language: 'ja',
-          timezone: 'Asia/Tokyo',
-        },
-        security: {
-          mfaRequired: false,
-          sessionTimeoutMinutes: 2, // min is 5
-          maxLoginAttempts: 5,
-        },
-        notifications: {
-          emailNotificationsEnabled: true,
-          systemAlertsEnabled: true,
-          maintenanceNotificationsEnabled: true,
-        },
-        appearance: {
-          theme: 'system',
-        },
-      };
-
-      const res = await app.request('/api/settings', {
-        method: 'PUT',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify(invalidSettings),
-      });
-
-      expect(res.status).toBe(400);
-      const body = await res.json();
-      expect(body.error).toBe('Validation error');
-    });
-  });
-
-  describe('GET /api/activities', () => {
-    it('最近のアクティビティを取得できるべき', async () => {
-      const mockLogs = [
-        {
-          id: '01HJXK5K3VDXK5YPNZBKRT5001',
-          tenantId: '',
-          userId: 'user-123',
-          action: 'CREATE',
-          resourceType: 'TENANT',
-          resourceId: 'tenant-001',
-          details: { name: 'Test Tenant' },
-          createdAt: new Date('2024-01-15T10:00:00.000Z'),
-        },
-        {
-          id: '01HJXK5K3VDXK5YPNZBKRT5002',
-          tenantId: '',
-          userId: 'user-456',
-          action: 'UPDATE',
-          resourceType: 'SETTING',
-          resourceId: 'setting-001',
-          details: {},
-          createdAt: new Date('2024-01-15T09:00:00.000Z'),
-        },
-      ];
-      mockAuditLogRepoFunctions.listByTenant.mockResolvedValue({
-        logs: mockLogs,
-        lastKey: undefined,
-      });
-
-      const res = await app.request('/api/activities');
-      expect(res.status).toBe(200);
-      const body = await res.json();
-      expect(body.data).toHaveLength(2);
-      expect(body.data[0]).toMatchObject({
-        id: '01HJXK5K3VDXK5YPNZBKRT5001',
-        action: 'CREATE',
-        resourceType: 'TENANT',
-        resourceId: 'tenant-001',
-        details: { name: 'Test Tenant' },
-      });
-      expect(body.pagination).toEqual({
-        limit: 10,
-        hasNextPage: false,
-      });
-    });
-
-    it('空のアクティビティリストを返すべき', async () => {
-      mockAuditLogRepoFunctions.listByTenant.mockResolvedValue({
-        logs: [],
-        lastKey: undefined,
-      });
-
-      const res = await app.request('/api/activities');
-      expect(res.status).toBe(200);
-      const body = await res.json();
-      expect(body.data).toEqual([]);
-      expect(body.pagination).toEqual({
-        limit: 10,
-        hasNextPage: false,
-      });
-    });
-
-    it('カスタム limit パラメータが動作すべき', async () => {
-      mockAuditLogRepoFunctions.listByTenant.mockResolvedValue({
-        logs: [],
-        lastKey: undefined,
-      });
-
-      const res = await app.request('/api/activities?limit=5');
-      expect(res.status).toBe(200);
-      const body = await res.json();
-      expect(body.pagination.limit).toBe(5);
-      expect(mockAuditLogRepoFunctions.listByTenant).toHaveBeenCalledWith('', {
-        limit: 5,
-      });
-    });
-
-    it('limit は 50 に制限されるべき', async () => {
-      mockAuditLogRepoFunctions.listByTenant.mockResolvedValue({
-        logs: [],
-        lastKey: undefined,
-      });
-
-      const res = await app.request('/api/activities?limit=999');
-      expect(res.status).toBe(200);
-      const body = await res.json();
-      expect(body.pagination.limit).toBe(50);
-    });
-
-    it('不正な limit は 10 にフォールバックすべき', async () => {
-      mockAuditLogRepoFunctions.listByTenant.mockResolvedValue({
-        logs: [],
-        lastKey: undefined,
-      });
-
-      const res = await app.request('/api/activities?limit=invalid');
-      expect(res.status).toBe(200);
-      const body = await res.json();
-      expect(body.pagination.limit).toBe(10);
-    });
-
-    it('hasNextPage が正しく設定されるべき', async () => {
-      mockAuditLogRepoFunctions.listByTenant.mockResolvedValue({
-        logs: [
-          {
-            id: '01HJXK5K3VDXK5YPNZBKRT5001',
-            tenantId: '',
-            userId: 'user-123',
-            action: 'CREATE',
-            resourceType: 'TENANT',
-            createdAt: new Date(),
-          },
-        ],
-        lastKey: { PK: 'AUDIT#', SK: 'LOG#2024-01-15T10:00:00.000Z#01HJXK' },
-      });
-
-      const res = await app.request('/api/activities?limit=1');
-      expect(res.status).toBe(200);
-      const body = await res.json();
-      expect(body.pagination.hasNextPage).toBe(true);
-    });
-
-    it('API エラー時は 500 エラーを返すべき', async () => {
-      mockAuditLogRepoFunctions.listByTenant.mockRejectedValue(
-        new Error('DynamoDB error')
-      );
-
-      const res = await app.request('/api/activities');
-      expect(res.status).toBe(500);
-      const body = await res.json();
-      expect(body.error).toBe('Failed to fetch activities');
-    });
-  });
-
-  describe('POST /api/tenants/:id/provision', () => {
-    it('PENDING ステータスでも backend 未設定時は 503 を返すべき', async () => {
-      const mockTenant = createMockTenant({ provisioningStatus: 'PENDING' });
-      mockTenantRepoFunctions.findById.mockResolvedValue(mockTenant);
-
-      const res = await app.request(`/api/tenants/${mockTenant.id}/provision`, {
-        method: 'POST',
-      });
-
-      expect(res.status).toBe(503);
-      const body = await res.json();
-      expect(body.error).toBe('Provisioning is not configured in this environment');
-    });
-
-    it('FAILED ステータスでも backend 未設定時は 503 を返すべき', async () => {
-      const mockTenant = createMockTenant({ provisioningStatus: 'FAILED' });
-      mockTenantRepoFunctions.findById.mockResolvedValue(mockTenant);
-
-      const res = await app.request(`/api/tenants/${mockTenant.id}/provision`, {
-        method: 'POST',
-      });
-
-      expect(res.status).toBe(503);
-      const body = await res.json();
-      expect(body.error).toBe('Provisioning is not configured in this environment');
-    });
-
-    it('IN_PROGRESS ステータスで 409 エラーになるべき', async () => {
-      const mockTenant = createMockTenant({
-        provisioningStatus: 'IN_PROGRESS',
-      });
-      mockTenantRepoFunctions.findById.mockResolvedValue(mockTenant);
-
-      const res = await app.request(`/api/tenants/${mockTenant.id}/provision`, {
-        method: 'POST',
-      });
-
-      expect(res.status).toBe(409);
-      const body = await res.json();
-      expect(body.error).toBe('Provisioning is already in progress');
-    });
-
-    it('COMPLETED ステータスで 409 エラーになるべき', async () => {
-      const mockTenant = createMockTenant({
-        provisioningStatus: 'COMPLETED',
-        applicationDeploymentStatus: 'DEPLOYED',
-      });
-      mockTenantRepoFunctions.findById.mockResolvedValue(mockTenant);
-
-      const res = await app.request(`/api/tenants/${mockTenant.id}/provision`, {
-        method: 'POST',
-      });
-
-      expect(res.status).toBe(409);
-      const body = await res.json();
-      expect(body.error).toBe('Tenant is already provisioned');
-    });
-
-    it('存在しないテナントで 404 エラーになるべき', async () => {
-      mockTenantRepoFunctions.findById.mockResolvedValue(null);
-
-      const nonExistentId = '01HJXK5K3VDXK5YPNZBKRT5XYZ';
-      const res = await app.request(`/api/tenants/${nonExistentId}/provision`, {
-        method: 'POST',
-      });
-
-      expect(res.status).toBe(404);
-      const body = await res.json();
-      expect(body.error).toBe('Tenant not found');
-    });
-
-    it('不正なID形式で 400 エラーになるべき', async () => {
-      const res = await app.request('/api/tenants/invalid@id!/provision', {
-        method: 'POST',
-      });
-
-      expect(res.status).toBe(400);
-      const body = await res.json();
-      expect(body.error).toBe('Invalid tenant ID');
-    });
-
-    it('backend 設定時は TenantOnboarding イベントを発行すべき', async () => {
-      vi.stubEnv('PROVISIONING_ENABLED', 'true');
-      const mockTenant = createMockTenant({ provisioningStatus: 'PENDING' });
-      mockTenantRepoFunctions.findById.mockResolvedValue(mockTenant);
-      mockTenantRepoFunctions.update.mockResolvedValue({
-        ...mockTenant,
-        provisioningStatus: 'IN_PROGRESS',
-        applicationDeploymentStatus: 'DEPLOYING',
-      });
-
-      const res = await app.request(`/api/tenants/${mockTenant.id}/provision`, {
-        method: 'POST',
-      });
-
-      expect(res.status).toBe(200);
-      expect(
-        mockProvisioningPublisherFunctions.publishTenantOnboarding
-      ).toHaveBeenCalledWith(mockTenant);
-      expect(mockTenantRepoFunctions.update).toHaveBeenCalledWith(mockTenant.id, {
-        provisioningStatus: 'IN_PROGRESS',
-        applicationDeploymentStatus: 'DEPLOYING',
-        provisioningError: null,
-      });
-    });
-
-    it('イベント発行失敗時は FAILED に戻すべき', async () => {
-      vi.stubEnv('PROVISIONING_ENABLED', 'true');
-      const mockTenant = createMockTenant({ provisioningStatus: 'PENDING' });
-      mockTenantRepoFunctions.findById.mockResolvedValue(mockTenant);
-      mockTenantRepoFunctions.update
-        .mockResolvedValueOnce({
-          ...mockTenant,
-          provisioningStatus: 'IN_PROGRESS',
-          applicationDeploymentStatus: 'DEPLOYING',
-        })
-        .mockResolvedValueOnce({
-          ...mockTenant,
-          provisioningStatus: 'FAILED',
-          applicationDeploymentStatus: 'FAILED',
-          provisioningError: 'EventBridge unavailable',
-        });
-      mockProvisioningPublisherFunctions.publishTenantOnboarding.mockRejectedValueOnce(
-        new Error('EventBridge unavailable')
-      );
-
-      const res = await app.request(`/api/tenants/${mockTenant.id}/provision`, {
-        method: 'POST',
-      });
-
-      expect(res.status).toBe(500);
-      expect(mockTenantRepoFunctions.update).toHaveBeenNthCalledWith(2, mockTenant.id, {
-        provisioningStatus: 'FAILED',
-        applicationDeploymentStatus: 'FAILED',
-        provisioningError: 'EventBridge unavailable',
-      });
-    });
-  });
-
-  describe('GET /api/tenants/:id/provision', () => {
-    it('プロビジョニングステータスを取得できるべき', async () => {
-      const mockTenant = createMockTenant({
-        provisioningStatus: 'COMPLETED',
-        applicationDeploymentStatus: 'NOT_DEPLOYED',
-        provisionedResources: { s3Prefix: 'tenants/test/' },
-        provisionedAt: new Date('2024-01-02T00:00:00.000Z'),
-      });
-      mockTenantRepoFunctions.findById.mockResolvedValue(mockTenant);
-
-      const res = await app.request(`/api/tenants/${mockTenant.id}/provision`);
-
-      expect(res.status).toBe(200);
-      const body = await res.json();
-      expect(body.tenantId).toBe(mockTenant.id);
-      expect(body.provisioningStatus).toBe('COMPLETED');
-      expect(body.applicationDeploymentStatus).toBe('NOT_DEPLOYED');
-      expect(body.provisionedResources).toEqual({ s3Prefix: 'tenants/test/' });
-      expect(body.provisionedAt).toBe('2024-01-02T00:00:00.000Z');
-      expect(body.provisioningEnabled).toBe(false); // Disabled in test env
-    });
-
-    it('存在しないテナントで 404 エラーになるべき', async () => {
-      mockTenantRepoFunctions.findById.mockResolvedValue(null);
-
-      const nonExistentId = '01HJXK5K3VDXK5YPNZBKRT5XYZ';
-      const res = await app.request(`/api/tenants/${nonExistentId}/provision`);
-
-      expect(res.status).toBe(404);
-      const body = await res.json();
-      expect(body.error).toBe('Tenant not found');
-    });
-
-    it('不正なID形式で 400 エラーになるべき', async () => {
-      const res = await app.request('/api/tenants/invalid@id!/provision');
-
-      expect(res.status).toBe(400);
-      const body = await res.json();
-      expect(body.error).toBe('Invalid tenant ID');
-    });
-  });
 });
